LayoutTests/inspector/canvas/recording-2d-saves-expected.txt

Test that CanvasManager is able to record actions made to 2D canvas contexts.


== Running test suite: Canvas.recording2D
-- Running test case: Canvas.recording2D.ExistingSaves
PASS: There should be 4 existing states.
PASS: State 0 should match expected fillStyle value.
PASS: State 1 should match expected fillStyle value.
PASS: State 2 should match expected fillStyle value.
PASS: State 3 should match expected fillStyle value.
